requirements.txt
python-dotenv
numpy
pandas
langchain
langchain-cohere
langchain-community
cohere
unstructured
chromadb
cohere
